# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Zia Carbon Footprint API",
        "description": "API documentation for the Zia Carbon Footprint Dashboard and Backend logic.",
        "contact": {
            "email": "support@zia.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Dynamic Host"
        }
    ],
    "paths": {
...
        "/api/periods/{period}/emissions": {
            "get": {
                "tags": [
                    "Emissions"
                ],
                "summary": "List all emissions for a specific period",
                "operationId": "650b1174696d77a7e85a6f8ba9531462",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of emissions",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Emissions"
                ],
                "summary": "Calculate and store a new carbon emission record",
                "operationId": "b686da77f5edd8b74baacdf606956fb0",
                "parameters": [
                    {
                        "name": "period",
                        "in": "path",
                        "description": "The ID of the period",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "emission_factor_id"
                                ],
                                "properties": {
                                    "emission_factor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "description": "Total activity data if monthly_inputs not provided",
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    },
                                    "monthly_inputs": {
                                        "description": "Array of monthly values for uncertainty analysis",
                                        "type": "array",
                                        "items": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "example": [
                                            10,
                                            20,
                                            30
                                        ]
                                    },
                                    "notes": {
                                        "type": "string",
                                        "example": "Monthly fuel consumption"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Emission record created and calculated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "calculated_co2e": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 5.67
                                        },
                                        "uncertainty_result": {
                                            "type": "number",
                                            "format": "float",
                                            "example": 0.25
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/emissions/{emission}": {
            "delete": {
                "tags": [
                    "Emissions"
                ],
                "summary": "Delete an emission record",
                "operationId": "88f8bf4106bce038f1744da40cfc3b23",
                "parameters": [
                    {
                        "name": "emission",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/companies": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "List all companies",
                "operationId": "a261a49d113a2af2a5ccb5798c6aae7e",
                "responses": {
                    "200": {
                        "description": "List of companies",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/companies/{company}/periods": {
            "get": {
                "tags": [
                    "Companies"
                ],
                "summary": "Get periods for a specific company",
                "operationId": "629b1e3b8bfbdad0f5eee9d6914d59e6",
                "parameters": [
                    {
                        "name": "company",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of periods",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/dictionaries/factors": {
            "get": {
                "tags": [
                    "Master Data"
                ],
                "summary": "Get all emission factors grouped by category",
                "description": "Use this to populate dropdowns in the frontend. Includes nested factors for each category.",
                "operationId": "1066c5ec0b7121723b7d5d5fdf9b7d86",
                "responses": {
                    "200": {
                        "description": "Grouped categories and factors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}